WDeclarativeMouseArea: Fix the 'wheel' event

+++ src/SkGui/src/declarative/WDeclarativeMouseArea.cpp
@@ -1339,7 +1339,7 @@ bool WDeclarativeMouseArea::sendMouseEvent(QMouseEvent * event)
 
     if (d->wheelEnabled)
     {
-        int steps = event->delta() / 120;
+        qreal steps = (qreal) (event->delta()) / 120;
 
         emit wheeled(steps);
     }
